--- grpcpack/pb/chainmsg.proto
@@ -1,41 +1,0 @@
-//protoc --go_out=plugins=grpc:. chainmsg.proto
-syntax = "proto3";
-package rpc.btc;
-
-service BtcService {
-    rpc GetNewAddress (AddressRequest) returns (AddressReply) {}
-    rpc CreateMultiSig (CreateMultiSigRequest) returns (AddressReply) {}
-    rpc GetBlockCount (EmptyRequest) returns (BlockCountReply) {}
-    rpc GetMiningInfo (EmptyRequest) returns (MiningInfoReply) {}
-}
-
-message CreateMultiSigRequest{
-    int32 minSignNum=1;
-    repeated string keys=2;
-}
-message AddressRequest {
-    string label = 1;
-}
-
-message AddressReply {
-    string address = 1;
-}
-
-message EmptyRequest {
-}
-
-message BlockCountReply {
-    int32 count = 1;
-}
-message MiningInfoReply {
-    int64 blocks = 1;
-    int64 currentblocksize = 2;
-    int64 currentblockweight = 3;
-    int64 currentblocktx = 4;
-    float difficulty = 5;
-    float networkhashps = 6;
-    int32 pooledtx = 7;
-    bool testnet = 8;
-    string chain = 9;
-}
-
