+++ service/schedule.go
@@ -19,11 +19,8 @@ var ScheduleService = scheduleManager{}
 
 func (service *scheduleManager) StartSchedule() {
 	go func() {
-		ticker10m := time.NewTicker(10 * time.Minute)
+		ticker10m := time.NewTicker(1 * time.Minute)
 		defer ticker10m.Stop()
-
-		//ticker := time.NewTicker(10 * time.Hour)
-		//defer ticker.Stop()
 
 		for {
 			select {
@@ -162,27 +159,6 @@ func checkRsmcAndSendBR(db storm.Node) {
 						}
 					}
 				}
-			}
-		}
-	}
-}
-
-func sendRdTx() {
-	var nodes []dao.RDTxWaitingSend
-	err := obdGlobalDB.Select(q.Eq("IsEnable", true)).Find(&nodes)
-	if err != nil {
-		return
-	}
-
-	for _, node := range nodes {
-		if tool.CheckIsString(&node.TransactionHex) {
-			_, err = rpcClient.SendRawTransaction(node.TransactionHex)
-			if err == nil {
-				if node.Type == 1 {
-					_ = addHTRD1aTxToWaitDB(node.HtnxIdAndHtnxRdId)
-				}
-				_ = obdGlobalDB.UpdateField(&node, "IsEnable", false)
-				_ = obdGlobalDB.UpdateField(&node, "FinishAt", time.Now())
 			}
 		}
 	}
